Move every class to it's own file
